user authorization

+++ app/Http/Controllers/ListingController.php
@@ -60,6 +60,10 @@ class ListingController extends Controller
     // Update listing Data
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort('403', 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -81,6 +85,9 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort('403', 'Unauthorized Action');
+        }
         // 删除操作
         $listing->delete();
